Uninstall the `users-permissions` plugin
This plugin was installed with the scaffolded project, but is not needed.
The plugin would be used to register end users to login and access content
from the system. Since the frontend for this project will be public, the plugin
won't be used and could cause confusion with admin users.