Update media uploads to account for multiple users
Instead of splitting the media files evenly among users, this update will
choose a random directory of files to upload. This change allows us to
not have to worry how many users may be in a given environment. The
users query is also updated to ensure we are only getting active users.